requirements/base.txt
django==3.2.8
django-model-utils==4.2.0
django-environ==0.8.0
djangorestframework==3.12.4
django-cors-headers==3.10.0
psycopg2==2.9.1